[ADD](Game_loop): Implement game state checks for player limits, collisions, and alive status
