+++ server/includes/server.h
@@ -133,4 +133,13 @@ void print_debug_info_connection(server_t *server, char *context);
 void print_debug_all(server_t *server, char *context, char *payload,
     unsigned char *header);
 
+// In game functions
+void check_jetpack(client_t *client, server_t *server);
+void check_limits(client_t *client);
+void check_entities_colisions(client_t *client, server_t *server);
+bool check_alive_begin(server_t *server, int *alive_count,
+    uint8_t *alive_player_id);
+void check_alive_end(client_t *client, int *alive_count,
+    uint8_t *alive_player_id, int i);
+
 #endif /* !SERVER_H_ */
